+++ src/main/java/com/example/kochbuch/controller/DetailController.java
@@ -1,0 +1,4 @@
+package com.example.kochbuch.controller;
+
+public class DetailController {
+}
