+++ src/main/java/com/example/kochbuch/controller/ListController.java
@@ -1,0 +1,4 @@
+package com.example.kochbuch.controller;
+
+public class ListController {
+}
