+++ src/main/java/com/example/kochbuch/controller/LoginController.java
@@ -1,0 +1,4 @@
+package com.example.kochbuch.controller;
+
+public class LoginController {
+}
