+++ src/main/java/com/example/kochbuch/model/ListModel.java
@@ -1,0 +1,4 @@
+package com.example.kochbuch.model;
+
+public class ListModel {
+}
