--- src/main/java/controller/DetailController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class DetailController {
-}
