--- src/main/java/controller/EditController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class EditController {
-}
