--- src/main/java/controller/ListController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class ListController {
-}
