--- src/main/java/controller/LoginController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class LoginController {
-}
