--- src/main/java/controller/UserController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class UserController {
-}
