--- src/main/java/model/DetailModel.java
@@ -1,4 +1,0 @@
-package model;
-
-public class DetailModel {
-}
