--- src/main/java/model/EditModel.java
@@ -1,4 +1,0 @@
-package model;
-
-public class EditModel {
-}
