--- src/main/java/model/ListModel.java
@@ -1,4 +1,0 @@
-package model;
-
-public class ListModel {
-}
