--- src/main/java/model/LoginModel.java
@@ -1,4 +1,0 @@
-package model;
-
-public class LoginModel {
-}
