+++ src/main/java/module-info.java
@@ -5,6 +5,6 @@ module com.example.kochbuch {
 
     opens com.example.kochbuch to javafx.fxml;
     exports com.example.kochbuch;
-    exports controller;
-    opens controller to javafx.fxml;
+    exports com.example.kochbuch.controller;
+    opens com.example.kochbuch.controller to javafx.fxml;
 }
