+++ src/main/resources/com/example/kochbuch/edit-view.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="controller.EditController"
+            fx:controller="com.example.kochbuch.controller.EditController"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
