+++ src/main/resources/com/example/kochbuch/list-view.fxml
@@ -1,14 +1,130 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.RowConstraints?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
 
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="controller.ListController"
-            prefHeight="400.0" prefWidth="600.0">
-
+<AnchorPane prefHeight="800.0" prefWidth="1200.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.kochbuch.controller.ListController">
+   <children>
+      <VBox layoutX="2.0" layoutY="-3.0" prefHeight="803.0" prefWidth="243.0" style="-fx-background-color: #D8DFDB;">
+         <children>
+            <ImageView fitHeight="50.0" fitWidth="156.0" pickOnBounds="true" preserveRatio="true" />
+            <Label text="KOCHBUCH">
+               <font>
+                  <Font name="Arial Black" size="21.0" />
+               </font>
+               <VBox.margin>
+                  <Insets left="50.0" />
+               </VBox.margin>
+            </Label>
+            <ImageView fitHeight="50.0" fitWidth="285.0" pickOnBounds="true" preserveRatio="true" />
+            <HBox prefHeight="19.0" prefWidth="378.0">
+               <children>
+                  <ImageView fitHeight="33.0" fitWidth="27.0" pickOnBounds="true" preserveRatio="true">
+                     <image>
+                        <Image url="@../../../images/icons/receipt-solid.jpg" />
+                     </image>
+                     <HBox.margin>
+                        <Insets left="100.0" />
+                     </HBox.margin>
+                  </ImageView>
+                  <Label prefHeight="33.0" prefWidth="102.0" text="Rezepte">
+                     <font>
+                        <Font name="Arial Bold" size="23.0" />
+                     </font>
+                     <HBox.margin>
+                        <Insets left="10.0" />
+                     </HBox.margin>
+                  </Label>
+               </children>
+            </HBox>
+            <ImageView fitHeight="50.0" fitWidth="244.0" pickOnBounds="true" preserveRatio="true" />
+            <HBox prefHeight="19.0" prefWidth="378.0">
+               <children>
+                  <ImageView fitHeight="33.0" fitWidth="35.0" pickOnBounds="true" preserveRatio="true">
+                     <image>
+                        <Image url="@../../../images/icons/right-to-bracket-solid.png" />
+                     </image>
+                     <HBox.margin>
+                        <Insets left="100.0" />
+                     </HBox.margin>
+                  </ImageView>
+                  <Label prefHeight="33.0" prefWidth="157.0" text="Log in">
+                     <font>
+                        <Font name="Arial Bold" size="23.0" />
+                     </font>
+                     <HBox.margin>
+                        <Insets left="10.0" />
+                     </HBox.margin>
+                  </Label>
+               </children>
+            </HBox>
+            <ImageView fitHeight="50.0" fitWidth="284.0" pickOnBounds="true" preserveRatio="true" />
+            <HBox prefHeight="40.0" prefWidth="378.0">
+               <children>
+                  <HBox prefHeight="19.0" prefWidth="378.0" />
+               </children>
+            </HBox>
+         </children>
+      </VBox>
+      <Label layoutX="1047.0" layoutY="26.0" prefHeight="38.0" prefWidth="79.0" />
+      <Label layoutX="1005.0" layoutY="756.0" text="by Arianit, Landry &amp; Enrico" />
+      <ScrollPane layoutX="317.0" layoutY="104.0" prefHeight="613.0" prefWidth="852.0" styleClass="scrollPane">
+         <content>
+            <GridPane prefHeight="657.0" prefWidth="770.0">
+              <columnConstraints>
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+              </columnConstraints>
+              <rowConstraints>
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+              </rowConstraints>
+               <children>
+                  <HBox prefHeight="100.0" prefWidth="200.0" style="-fx-background-color: #D8DFDB;" styleClass="recipeOverview">
+                     <children>
+                        <ImageView fitHeight="231.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true">
+                           <image>
+                              <Image url="@../../../images/RezeptBilder/spaghetti.png" />
+                           </image>
+                           <HBox.margin>
+                              <Insets top="50.0" />
+                           </HBox.margin>
+                        </ImageView>
+                        <VBox prefHeight="231.0" prefWidth="193.0">
+                           <children>
+                              <VBox prefHeight="231.0" prefWidth="185.0">
+                                 <children>
+                                    <Label text="Spaghetti Pomodoro">
+                                       <font>
+                                          <Font name="Arial Bold" size="18.0" />
+                                       </font>
+                                       <VBox.margin>
+                                          <Insets top="10.0" />
+                                       </VBox.margin>
+                                    </Label>
+                                 </children>
+                              </VBox>
+                           </children>
+                        </VBox>
+                     </children>
+                  </HBox>
+               </children>
+            </GridPane>
+         </content>
+         <padding>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </padding>
+      </ScrollPane>
+   </children>
 </AnchorPane>
