+++ src/main/resources/com/example/kochbuch/login-view.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="controller.LoginController"
+            fx:controller="com.example.kochbuch.controller.LoginController"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
