+++ src/main/resources/com/example/kochbuch/styles.css
@@ -1,0 +1,8 @@
+.recipeOverview {
+    -fx-background-radius: 10px;
+    -fx-effect: dropshadow(gaussian, rgba(0,0,0,0.5), 10, 0, 0, 0);
+}
+
+.scrollPane {
+    -fx-border-width: 0px;
+    }
