+++ src/main/resources/com/example/kochbuch/user-view.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="controller.UserController"
+            fx:controller="com.example.kochbuch.controller.UserController"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
